Review how word embeddings match by case
