+++ src/main/scala/com/johnsnowlabs/nlp/embeddings/EmbeddingsReadable.scala
@@ -3,7 +3,7 @@ package com.johnsnowlabs.nlp.embeddings
 import com.johnsnowlabs.nlp.ParamsAndFeaturesReadable
 import org.apache.spark.sql.SparkSession
 
-trait EmbeddingsReadable[T <: WordEmbeddingsModel] extends ParamsAndFeaturesReadable[T] {
+trait EmbeddingsReadable[T <: WordEmbeddings] extends ParamsAndFeaturesReadable[T] {
   def readEmbeddings(instance: T, path: String, spark: SparkSession): Unit = {
     instance.deserializeEmbeddings(path, spark)
   }
