+++ src/test/scala/com/johnsnowlabs/nlp/AnnotatorBuilder.scala
@@ -10,7 +10,7 @@ import com.johnsnowlabs.nlp.annotators.sda.pragmatic.SentimentDetector
 import com.johnsnowlabs.nlp.annotators.sda.vivekn.ViveknSentimentApproach
 import com.johnsnowlabs.nlp.annotators.spell.norvig.NorvigSweetingApproach
 import com.johnsnowlabs.nlp.training.POS
-import com.johnsnowlabs.nlp.embeddings.{WordEmbeddings, WordEmbeddingsFormat, WordEmbeddingsModel}
+import com.johnsnowlabs.nlp.embeddings.{WordEmbeddings, WordEmbeddingsFormat}
 import com.johnsnowlabs.nlp.util.io.{ExternalResource, ReadAs, ResourceHelper}
 import org.apache.spark.ml.Pipeline
 import org.apache.spark.sql.{Dataset, Row}
@@ -235,14 +235,11 @@ object AnnotatorBuilder extends FlatSpec { this: Suite =>
     getGLoveEmbeddings(dataset).transform(df)
   }
 
-  def getGLoveEmbeddings(dataset: Dataset[Row]): WordEmbeddingsModel = {
-    val df = withFullPOSTagger(dataset)
-
+  def getGLoveEmbeddings(dataset: Dataset[Row]): WordEmbeddings = {
     new WordEmbeddings()
       .setEmbeddingsSource("src/test/resources/ner-corpus/embeddings.100d.test.txt", 100, WordEmbeddingsFormat.TEXT)
       .setInputCols("sentence", "token")
       .setOutputCol("embeddings")
-      .fit(df)
   }
 
   def withNerDLTagger(dataset: Dataset[Row]): Dataset[Row] = {
